% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculators.R
\name{as.quantile}
\alias{as.quantile}
\title{Quantile Transformation}
\usage{
as.quantile(x, ...)
}
\arguments{
\item{x}{The input vector}

\item{...}{(\code{\link[rlang]{dots_list}}): Additional arguments sent to \code{\link[stats]{quantile}}}
}
\value{
A quantile representation of the input
}
\description{
\code{as.quantile} is a wrapper for \code{\link[stats]{quantile}}.
}
\seealso{
Other Chapter 1 - Calculators: 
\code{\link{calc.geo_mean}()},
\code{\link{calc.harmonic_mean}()},
\code{\link{calc.means}()},
\code{\link{calc.rms}()},
\code{\link{calc.zero_mean}()},
\code{\link{range_diff}()},
\code{\link{ranking.algorithm}()},
\code{\link{ratio}()}
}
\concept{Chapter 1 - Calculators}
