% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{as.recursive}
\alias{as.recursive}
\title{Recast a Function as Recursive}
\usage{
as.recursive(fun, cond_def, finalize = I)
}
\arguments{
\item{fun}{A function}

\item{cond_def}{A lambda expression defining the condition for which recursion continues}

\item{finalize}{A lambda expression or function that finalizes the result}
}
\description{
\code{as.recursive} creates a recursive version of the function passed to \code{fun}.
}
