% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.of.utilities.R
\name{as.regex}
\alias{as.regex}
\title{Tag a Regular-Expression String}
\usage{
as.regex(...)
}
\arguments{
\item{...}{The string to set "as regex"}
}
\value{
The string appended with attribute "regex" set to \code{TRUE}
}
\description{
\code{as.regex} adds class "regex" to a strings passed to \code{...}
}
