% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculators.R
\name{as_quantile}
\alias{as_quantile}
\title{Quantiles Transformation}
\usage{
as_quantile(x, ...)
}
\arguments{
\item{x}{The input vector}

\item{...}{(\code{\link[rlang]{dots_list}}): Additional arguments sent to \code{\link[stats]{quantile}}}
}
\value{
A quantile representation of the input
}
\description{
\code{as_quantile} is a wrapper for \code{\link[stats]{quantile}}.
}
\seealso{
Other Calculators: 
\code{\link{range_diff}()},
\code{\link{ranking.algorithm}()},
\code{\link{ratio}()}
}
\concept{Calculators}
