% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0. description.R
\name{book.of.utilities-package}
\alias{book.of.utilities-package}
\title{Utilities for Working with Data}
\description{
\code{book.of.utilities} provides utility functions to help with data-manipulation, mathematical and relational operational tasks.  The suite of functions are divided into the following families:

\itemize{
\item{Calculators: }
\item{Object Management: }
\item{Counters: }
\item{Custom Operators: }
\item{Miscellaneous: }
}
}
