% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculators.R
\name{calc.harmonic_mean}
\alias{calc.harmonic_mean}
\title{Calculate the Harmonic Mean}
\usage{
calc.harmonic_mean(a, post.op = eval)
}
\arguments{
\item{a}{(vector) A vector of numeric values}

\item{post.op}{See \code{\link{calc.means}}}
}
\description{
This is an alias that calls \code{\link{calc.means}(a, "hm")}
}
\seealso{
Other Calculators: 
\code{\link{calc.geo_mean}()},
\code{\link{calc.means}()},
\code{\link{calc.rms}()},
\code{\link{calc.zero_mean}()},
\code{\link{radix}()},
\code{\link{range_diff}()},
\code{\link{ranking.algorithm}()},
\code{\link{ratio}()}
}
\concept{Calculators}
