% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.of.utilities.R
\name{calc.range_diff}
\alias{calc.range_diff}
\title{Magnitude of a Range of Values}
\usage{
calc.range_diff(...)
}
\arguments{
\item{...}{(numeric[]): A numeric vector for which the range and corresponding difference is calculated. If a list is provided, the calculation is executed over each element of the list.}
}
\value{
A scalar value of the inclusive difference of upper and lower boundaries of the range of the input vector
}
\description{
\code{calc.range_diff} calculates the difference in the range of values given by \code{input}
}
\seealso{
Other Calculators: 
\code{\link{ranking.algorithm}()},
\code{\link{ratio}()}
}
\concept{Calculators}
