% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.of.utilities.R
\name{calc.zero_mean}
\alias{calc.zero_mean}
\title{Calculate the Zero-Mean}
\usage{
calc.zero_mean(a, post.op = eval)
}
\arguments{
\item{a}{(vector) A vector of numeric values}

\item{post.op}{See \code{\link{calc.means}}}
}
\description{
\code{calc.zero_mean} subtracts the mean from the input
}
\details{
This is an alias that calls \code{\link{calc.means}(a, "hm")}
}
\seealso{
Other Central-tendency calculations: 
\code{\link{calc.geo_mean}()},
\code{\link{calc.harmonic_mean}()},
\code{\link{calc.means}()},
\code{\link{calc.rms}()}
}
\concept{Central-tendency calculations}
