% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{checksum}
\alias{checksum}
\title{Checksum Validation}
\usage{
checksum(object, hash, ...)
}
\arguments{
\item{object}{See \code{\link[digest]{digest}}}

\item{hash}{(string) The hash to compare}

\item{...}{See \code{\link[digest]{digest}}}
}
\value{
A logical scalar
}
\description{
\code{checksum} provides a wrapper to \code{\link[digest]{digest}} to make checksum validation easy
}
