% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counters.R
\name{count.cycles}
\alias{count.cycles}
\title{Cycle Counter}
\usage{
count.cycles(cond, offset = 1, reset)
}
\arguments{
\item{cond}{(logical[]) A logical vector}

\item{offset}{(integer|1) An integer defining the series origin}

\item{reset}{(logical[]) A logical vector of the same length as \code{cond} such that when \code{TRUE}, the cycle counter resets to the offset. Un-evaluated expressions are also accepted and should refer to \code{cond} in the definition.}
}
\value{
A one-based (default), conditionally-incremented series
}
\description{
\code{count.cycles} detects and maps conditional occurrence frequency
}
\examples{
sample(c(TRUE, FALSE), 50, TRUE) |>
print() \%>\%
count.cycles(reset = cumsum(.) \%\% 5 == 0)

}
\seealso{
Other Counters: 
\code{\link{factor.int}()}
}
\concept{Counters}
