% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.of.utilities.R
\name{do_count}
\alias{do_count}
\title{Count Stuff}
\usage{
do_count(i, distinct = FALSE)
}
\arguments{
\item{i}{A vector of things to count}

\item{distinct}{(logical) When \code{TRUE}, a distinct count of objects is returned}
}
\value{
The count of elements in \code{i}, possibly distinct
}
\description{
Count Stuff
}
\seealso{
Other Counters: 
\code{\link{count.cycles}()},
\code{\link{factor.int}()}
}
\concept{Counters}
