% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3. counters.R
\name{factor.int}
\alias{factor.int}
\title{Factorization of Integers}
\usage{
factor.int(i, ...)
}
\arguments{
\item{i}{(integer[]) One or more integers to factor}

\item{...}{Additional integers to factor (can be used in conjunction with \code{i})}
}
\value{
The factors of the input given as an vector or list of vectors
}
\description{
Factorization of Integers
}
\seealso{
Other Counters: 
\code{\link{count.cycles}()},
\code{\link{do_count}()}
}
\concept{Counters}
