% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2. object_management.R
\name{fill.na}
\alias{fill.na}
\title{Fill in a Sparse Vector
\code{fill.na} fills in positions equivalent to \code{val} with the first preceding value not equal to \code{val}}
\usage{
fill.na(i, incl.null = TRUE)
}
\arguments{
\item{i}{The input vector}

\item{incl.null}{(logical) Should \code{NULL} values be accepted with \code{NA}'s?}
}
\value{
The input vector with values filled in
}
\description{
Fill in a Sparse Vector
\code{fill.na} fills in positions equivalent to \code{val} with the first preceding value not equal to \code{val}
}
\seealso{
Other Object management: 
\code{\link{distinct.list}()},
\code{\link{enlist}()},
\code{\link{get.object_sizes}()},
\code{\link{scrub.data}()}
}
\concept{Object management}
