% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{gen_pass}
\alias{gen_pass}
\title{Generate a Password
Use \code{gen_pass} to generate a password and automatically save it to the clipboard to paste.  The \code{random} package is the engine for generating output when an internet connection is available; otherwise, \code{\link[base]{sample}} is used.}
\usage{
gen_pass(
  out_length = 24,
  .symbols = c("!", "@", "#", "$", "\%", ")", "(", "."),
  .alpha = c(letters, LETTERS),
  .nums = 1:9
)
}
\arguments{
\item{out_length}{(integer) The length of the output}

\item{.alpha, .nums, .symbols}{Letters, numbers, and symbols to use}
}
\value{
Invisibly, a string of length \code{out_length}
}
\description{
Generate a Password
Use \code{gen_pass} to generate a password and automatically save it to the clipboard to paste.  The \code{random} package is the engine for generating output when an internet connection is available; otherwise, \code{\link[base]{sample}} is used.
}
