% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.of.utilities.R
\name{\%bin\%}
\alias{\%bin\%}
\title{Fixed-Interval Binning Operator}
\usage{
x \%bin\% b
}
\arguments{
\item{x}{(vector) A vector of numeric values}

\item{b}{(integer) A bin size to use}
}
\value{
The input after subtracting Modulo(b): x - (x \%\% b)
}
\description{
\code{\%bin\%} creates a vector of values in \code{a} represented as multiples of \code{b}
}
\seealso{
Other Custom operators: 
\code{\link{\%::\%}()},
\code{\link{\%><\%}()},
\code{\link{\%??\%}()},
\code{\link{\%?\%}()},
\code{\link{\%all\%}()}
}
\concept{Custom operators}
