% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.of.utilities.R
\name{\%::\%}
\alias{\%::\%}
\title{Decision Operator}
\usage{
tr \%::\% fls
}
\arguments{
\item{tr}{(logical) : Scalar or vector value to return as the TRUE  value}

\item{fls}{(logical) : Scalar or vector value to return as the FALSE value}

\item{id}{(scalar) : Scalar or vector value to return as the identifier for the result}
}
\value{
a \code{\link[data.table]{data.table}} of resultant values options for \code{true} and \code{false}
}
\description{
Use \code{\%::\%} anywhere where binary choice vectors can be useful
}
\seealso{
Other Custom operators: 
\code{\link{\%><\%}()},
\code{\link{\%??\%}()},
\code{\link{\%?\%}()},
\code{\link{\%all\%}()},
\code{\link{\%bin\%}()}
}
\concept{Custom operators}
