% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_operators.R
\name{\%><\%}
\alias{\%><\%}
\title{\code{\link{test_between}} Operator}
\usage{
a \%><\% z
}
\arguments{
\item{a}{Preferably a data table or data frame, but any object that can be coerced into such}

\item{z}{Preferably a data table or data frame, but any object that can be coerced into such}
}
\value{
An Boolean object comprised of the results of testing for "between-ness"
}
\description{
\code{\link{test_between}} Operator
}
\seealso{
Other Custom operators: 
\code{\link{\%??\%}()},
\code{\link{\%?\%}()},
\code{\link{\%bin\%}()},
\code{\link{\%tf\%}()},
\code{\link{test_between}()}
}
\concept{Custom operators}
