% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_operators.R
\name{\%?\%}
\alias{\%?\%}
\title{Simple IF-THEN-ELSE Operator}
\usage{
cond \%?\% result
}
\arguments{
\item{cond}{(logical) A \emph{vector} that evaluates to \code{TRUE} or \code{FALSE}}

\item{result}{(tensor) Resultant values for TRUE and FALSE conditionals}
}
\value{
A \code{\link[data.table]{data.table}} object comprised of values occupying the 'then' and 'else' slots in the 'if-then-else' logical test
}
\description{
When \code{cond} is \code{TRUE}, \code{result$true} is returned, and the same for \code{cond == FALSE}.  For \code{result}, the easiest way to set the available choices is to use \code{\%::\%}; otherwise, a environment(-like) object with members named 'true' and 'false' must be provided
}
\seealso{
Other Custom Operators: 
\code{\link{\%><\%}()},
\code{\link{\%??\%}()},
\code{\link{\%bin\%}()},
\code{\link{\%tf\%}()},
\code{\link{test_between}()}
}
\concept{Custom Operators}
