% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_operators.R
\name{\%if_na_empty\%}
\alias{\%if_na_empty\%}
\title{Replace NA and Empty Values}
\usage{
x \%if_na_empty\% y
}
\arguments{
\item{x, y}{The input and replacement values, respectively. Data types are treated as per \code{\link[purrr]{modify2}}, and arguments must be of equal length}
}
\value{
The input with \code{NA} and empty values replaced.
}
\description{
This function is a wrapper for \code{\link[base]{ifelse}} by means of \code{\link[purrr]{modify2}}. \code{NA} and empty values in \code{x} are replaced by values provided in \code{y}.
}
\seealso{
Other Custom Operators: 
\code{\link{\%><\%}()},
\code{\link{\%??\%}()},
\code{\link{\%?\%}()},
\code{\link{\%bin\%}()},
\code{\link{\%tf\%}()},
\code{\link{test_between}()}
}
\concept{Custom Operators}
