% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.of.utilities.R
\name{is.regex}
\alias{is.regex}
\title{Test for Class 'regex'}
\usage{
is.regex(i)
}
\arguments{
\item{i}{(string[]) A string or string vector}
}
\description{
\code{is.regex} returns \code{TRUE} when the value of \code{i} contains class "regex"
}
