% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{keyring_export}
\alias{keyring_export}
\title{Export keyring Entries}
\usage{
keyring_export(keyring = NULL, as.raw = FALSE)
}
\arguments{
\item{keyring}{(string[]) The name(s) of keyrings to export (defaults to all named keyrings when calling \code{\link[keyring]{keyring_list}})}

\item{as.raw}{(logical | FALSE) Should each entry be cast as a raw vector?}
}
\value{
Keyring entries as JSON or raw-encoded JSON
}
\description{
\code{keyring_export} creates JSON output for available \code{\link[keyring]{keyring}}s
}
\seealso{
Other Authentication Functions: 
\code{\link{gen.pass}()},
\code{\link{kr_key}()}
}
\concept{Authentication Functions}
