% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{keyring_import}
\alias{keyring_import}
\title{Import keyring Entries}
\usage{
keyring_import(data, ...)
}
\arguments{
\item{data}{The named list of exported keyring data}

\item{...}{Additional named keyring entries}
}
\value{
Keyring entries as JSON or raw-encoded JSON
}
\description{
\code{keyring_import} registers exported \code{\link[keyring]{keyring}}s (see \code{\link{keyring_export}})
}
\note{
Unnamed inputs will not be imported but indicated via console message
}
\seealso{
Other keyring Utilities: 
\code{\link{keyring_export}()}
}
\concept{keyring Utilities}
