% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{kr_key}
\alias{kr_key}
\title{Store an encrypted \code{keyring} key}
\usage{
kr_key(service = character(0), username = character(0), keyring = character(0))
}
\description{
The \code{kr_key} class
}
\section{Slots}{

\describe{
\item{\code{service,usernane,keyring}}{See \code{\link[keyring]{key_get}}}

\item{\code{key,password}}{An anonymous function to retrieve and return the value of the current \code{keyring} key.}

\item{\code{update}}{An anonymous function to set a new value for the \code{keyring} key referenced by the object.}
}}

\seealso{
Other Authentication Functions: 
\code{\link{gen_pass}()},
\code{\link{keyring_export}()},
\code{\link{keyring_import}()}
}
\concept{Authentication Functions}
