% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{log_note}
\alias{log_note}
\title{Create an Entry into a Development Log}
\usage{
log_note(..., file = "", show = FALSE, append = TRUE)
}
\arguments{
\item{...}{(\code{\link[rlang]{dots_list}}): The string(s) of text to write to the file.  Multiple values are collapsed into a single string separated by a newline. Expressions are allowed and will have the captured output included.}

\item{file}{(string|"") The name of the file to (over)write.  If no file is given, the output is sent to the console.}

\item{show}{(logical | \code{FALSE}) When \code{TRUE}, print the contents of \code{file} (requires library \code{readtext}).}

\item{append}{(logical | \code{TRUE}) When \code{TRUE} new content is appended.}
}
\value{
See parameter \code{show}
}
\description{
\code{log_note} serves as a wrapper for \code{\link[base]{cat}} and provides an easy way to write to a plain-text file
Even though the original use-case is the real-time capturing of developer notes, it can be used for incremental writing to a file when needed.
}
\seealso{
Other Miscellaneous Functions: 
\code{\link{as.regex}()},
\code{\link{checksum}()},
\code{\link{gen.primes}()},
\code{\link{is.regex}()},
\code{\link{polyname2orig}()},
\code{\link{unregex}()},
\code{\link{vlogical}()}
}
\concept{Miscellaneous Functions}
