% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.of.utilities.R
\name{mark.time}
\alias{mark.time}
\title{Time Stamp Utility}
\usage{
mark.time(
  msg = "",
  reset = TRUE,
  lap = FALSE,
  end = FALSE,
  history = FALSE,
  ...
)
}
\arguments{
\item{msg}{(string) The message to print along side the timestamp; defaults to \code{""}}

\item{reset}{(logical | TRUE) \code{TRUE} resets the closure and replaces the stopwatch}

\item{lap}{(logical | FALSE) \code{TRUE} preserves the closure and appends the stopwatch}

\item{end}{(logical | FALSE) \code{TRUE} resets closure and appends the stopwatch.}

\item{history}{(logical | FALSE) \code{TRUE} shows the current contents of the stopwatch and then exits without taking additional action}

\item{...}{Not used.}
}
\value{
A copy of the library function
}
\description{
\code{mark.time} generates timestamps during execution workflow
}
\details{
Assign this function to a variable as a delegate and invoke the variable as a function.  By default, each call to the function will toggle the message returned indicating the beginning or end of execution along with the time of invocation.  See \href{http://adv-r.had.co.nz/Functional-programming.html#closures}{Advanced R by Hadley Wickham: Closures} for additional context. All parameters are independent; therefore, the logical \code{AND} should be assumed
}
