% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.of.utilities.R
\name{polyname2orig}
\alias{polyname2orig}
\title{Polynomial Names to Original}
\usage{
polyname2orig(poly.names, orig.names, degree, ...)
}
\arguments{
\item{poly.names}{(string[]) The variable names after polynomial conversion}

\item{orig.names}{(string[]) The variable names before polynomial conversion}

\item{degree}{(int) The degree used in polynomial conversion}

\item{...}{Arguments to send to \code{\link[base]{paste}}, namely \code{sep} (default "_") and \code{collapse} (default ("\emph{x}"))}

\item{i}{An object created by the call to \code{\link[stats]{poly}}}
}
\value{
The polynomial names vector with positions replaced with original name elements
}
\description{
\code{polyname2orig} renames the column output of the call to \code{\link[stats]{poly}} by replacing each ordinal slot with the original name
}
