% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{radix}
\alias{radix}
\title{Radix Conversion}
\usage{
radix(x, ...)
}
\arguments{
\item{x}{(integer[]) A vector of integers.  Coercion for non-integer input will result in loss of precision.}

\item{...}{(string,symbol,integer) The target radix to which the values in \code{x} will be converted.\cr
\itemize{
\item{Length: If two values are provided, the output is the conversion from the first radix to the second}
\item{String/symbol: If given as a strings or symbols, the following are supported (partial matching supported):\cr
\itemize{
\item{\code{binary}}
\item{\code{octal}}
\item{\code{decimal}}
\item{\code{duodecimal}}
\item{\code{hexadecimal}}
\item{\code{vigesimal}}
\item{\code{sexagesimal}}
}
}
}}
}
\description{
Radix Conversion
}
\examples{
# As binary
radix(c(12, 20), bin)

# To hexadecimal
radix(c(12, 20), hex)

# As hex to decimal
radix(c(18, 32), hex, dec)

# Hex to binary vs. as hex to binary (equivalent outcomes)
radix(c(18, 32), hex, bin)
radix(c(12, 20), hex) |> radix(hex, bin)
}
