% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculators.R
\name{range_diff}
\alias{range_diff}
\title{Magnitude of a Range of Values}
\usage{
range_diff(...)
}
\arguments{
\item{...}{(\code{\link[rlang]{dots_list}}): A numeric vector for which the range and corresponding difference is calculated. If a list is provided, the calculation is executed over each element of the list.}
}
\value{
A scalar value of the inclusive difference of upper and lower boundaries of the range of the input vector
}
\description{
\code{range_diff} calculates the difference in the range of values given by \code{input}
}
\seealso{
Other Calculators: 
\code{\link{calc.geo_mean}()},
\code{\link{calc.harmonic_mean}()},
\code{\link{calc.means}()},
\code{\link{calc.rms}()},
\code{\link{calc.zero_mean}()},
\code{\link{odds2probs}()},
\code{\link{radix}()},
\code{\link{ranking.algorithm}()},
\code{\link{ratio}()}
}
\concept{Calculators}
