% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculators.R
\name{ranking.algorithm}
\alias{ranking.algorithm}
\title{Ranking Algorithm}
\usage{
ranking.algorithm(
  scores,
  rank.size,
  poss.scores,
  test = `>=`,
  rescale = 1,
  map.only = FALSE,
  show.all = FALSE,
  use.futures = FALSE,
  dflt.fn = min,
  map.fn = (purrr::map)
)
}
\arguments{
\item{scores}{(vector) The vector of numeric scores to rank}

\item{rank.size}{(integer) The number of ranks to allow}

\item{poss.scores}{(vector) The distinct values demarcating each rank threshold}

\item{test}{(object) A function or operator that takes in as input the scores, performs a conditional test, and returns a logical vector of equal length}

\item{rescale}{(numeric) A multiplier that scales the scores vector up or down}

\item{map.only}{(logical|FALSE) When \code{TRUE}, a \href{https://www.rdocumentation.org/packages/data.table/versions/1.12.8}{data.table} object of the rank and rank thresholds is returned}

\item{show.all}{(logical|FALSE) When \code{TRUE}, a \href{https://www.rdocumentation.org/packages/data.table/versions/1.12.8}{data.table} object of the scores, ranks thresholds, and ranks are returned}

\item{use.futures}{(logical|FALSE) When \code{TRUE}, the operation is parallelized using \href{https://www.rdocumentation.org/packages/future/versions/1.17.0}{future}s}

\item{dflt.fn}{(function) The default function to use when replacing \code{NA} values in ranks}

\item{map.fn}{(function) The function to iterate the scores when computing the output}
}
\description{
\code{ranking.algorithm} ranks values across a vector of values
}
\seealso{
Other Calculators: 
\code{\link{as_quantile}()},
\code{\link{range_diff}()},
\code{\link{ratio}()}
}
\concept{Calculators}
