% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculators.R
\name{ratio}
\alias{ratio}
\title{Ratio Calculator}
\usage{
ratio(
  i,
  type = "of.sum",
  as_density = FALSE,
  sort.type = c("none", "value", "label", "num_label"),
  ...
)
}
\arguments{
\item{i}{(vector) numeric vector}

\item{type}{(string[]) One or more types of ratio methods to use (see Details): a vector of supported values is supported}

\item{as_density}{(logical) \code{TRUE} returns \code{x * p(1 - p)}, where \code{p} is a vector of cumulative proportions of \code{x}}

\item{sort.type}{(string,function) One of the following:\cr
\enumerate{
\item{\code{"none"} to return along the original order (default)}
\item{\code{"value"} to sort by value}
\item{\code{"label"} to sort by name: all elements must be named}
\item{\code{"num_label"} to sort by numeral names: all elements must be named}
}\cr}

\item{...}{\code{\link[rlang]{dots_list}}: provided for backwards-compatibility as the function signature changes.
Sorting occurs \emph{before} the input is further processed.}
}
\value{
A numeric vector
}
\description{
\code{ratio} calculates one of the following ratio types:\cr
\enumerate{
\item{\code{"of.sum"} (relative to the sum of \code{i})}
\item{\code{"of.max"} (relative to maximum value)}
\item{\code{"cumulative"} (cumulative total vs. total)}
}.\cr Using the related operator \code{\%ratio\%} assumes simple division by the total of \code{i}.
}
\note{
Because this function produces values on a \emph{ratio} scale, all values are internally shifted such that all values are >= 0.
}
\examples{
 (x <- sample(-100:100, 30))

 (d <- cbind(
 	x
 	, of.sum = ratio(x, of.sum)
 	, of.sum.dens = ratio(x, of.sum, as = TRUE)
 	, of.max = ratio(x, of.max)
 	, of.max.dens = ratio(x, of.max, as = TRUE)
 	, cumulative = ratio(x, cumulative)
 	, cumulative.dens = ratio(x, cumulative, as = TRUE)
 	))

 plot(of.sum ~ x, data = d, main = "of.sum", col = "red")
 plot(of.max ~ x, data = d, main = "of.max", col = "green")
 plot(cumulative ~ x, data = d, main = "cumulative", col = "blue")
 plot(of.sum.dens ~ x, data = d, main = "of.sum w/ density", col = "red")
 points(of.max.dens ~ x, data = d, main = "of.max w/ density", col = "green")
 points(cumulative.dens ~ x, data = d, main = "cumulative w/ density", col = "blue")

 # sort.type:
 (x2 <- sample(100, 5, TRUE) \%>\% rlang::set_names(. + 50))

 list(
 	`orig := table(...)` = x2
 	, `ratio(<args>, sort.type = none)` = ratio(x2, sort.type = "none")
 	, `ratio(<args>, sort.type = value)` = ratio(x2, sort.type = "value")
 	, `ratio(<args>, sort.type = label)` = ratio(x2, sort.type = "label")
 	, `ratio(<args>, sort.type = num_label)` = ratio(x2, sort.type = "num_label")
 	)

}
\seealso{
Other Calculators: 
\code{\link{calc.geo_mean}()},
\code{\link{calc.harmonic_mean}()},
\code{\link{calc.means}()},
\code{\link{calc.rms}()},
\code{\link{calc.zero_mean}()},
\code{\link{odds2probs}()},
\code{\link{radix}()},
\code{\link{range_diff}()},
\code{\link{ranking.algorithm}()}
}
\concept{Calculators}
