% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.of.utilities.R
\name{scrub.data}
\alias{scrub.data}
\title{Scrub and Replace Values}
\usage{
scrub.data(input, condFn = is.na, replacement, ...)
}
\arguments{
\item{...}{Not used}

\item{input:}{An object that contains values to scrub}

\item{condFn:}{A function name or function expression that serves as the test for values to scrub}

\item{replacement:}{The replacement value}
}
\description{
\code{scrub.data} replaces values that meet the condition given by the function passed to \code{condFn}
}
\seealso{
Other Object management: 
\code{\link{distinct.list}()},
\code{\link{enlist}()},
\code{\link{fill.na}()}
}
\concept{Object management}
