% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{vlogical}
\alias{vlogical}
\title{Vectorized Logical Tests}
\usage{
vlogical(vector, vpattern, test, simplify_with = NULL, ...)
}
\arguments{
\item{vector}{A vector or dimensional object to be processed (e.g., matrix, data.frame, etc.)}

\item{vpattern}{A vector of patterns to be matched}

\item{test}{(function) The function to use for logical testing: the function should be appropriate for the values of \code{vpattern} and have a named argument of \code{vector}, an additional argument to accept the pattern fed atomically, and '...' (even if not used).}

\item{simplify_with}{(function) When provided, the function operates over rows, thereby simplifying the result}

\item{...}{Additional arguments to be sent to the function held by argument \code{test}}
}
\value{
A logical matrix, with rows of the same length as \code{vector} and columns the length of \code{vpattern} TRUE for items that match pattern. If vector is dimensional, multiple \code{TRUE} values may be found for each row.
}
\description{
\code{vlike} is a vectorized version of \code{\link[data.table]{like}} allowing a pattern vector to be supplied
}
\examples{
vlogical(
	vector = data.table::data.table(t(sapply(1:20, function(i){ c(a = sample(LETTERS, 1), b = sample(letters, 1)) })))
	, c(sample(LETTERS, 5), sample(letters, 5))
	, test = function(vector, q, ...){ any(unique(vector) \%in\% unique(q))}
	, simplify_with = mean
	, ignore.case = TRUE
)

}
\seealso{
Other Chapter 5 - Miscellaneous Functions: 
\code{\link{as.regex}()},
\code{\link{checksum}()},
\code{\link{gen.pass}()},
\code{\link{gen.primes}()},
\code{\link{is.regex}()},
\code{\link{log_note}()},
\code{\link{polyname2orig}()},
\code{\link{unregex}()}
}
\concept{Chapter 5 - Miscellaneous Functions}
