% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{as.regex}
\alias{as.regex}
\title{Tag a Regular-Expression String}
\usage{
as.regex(...)
}
\arguments{
\item{...}{(\code{\link[rlang]{dots_list}}): The string to set "as regex"}
}
\value{
The string appended with attribute "regex" set to \code{TRUE}
}
\description{
\code{as.regex()} adds class "regex" to a strings passed to \code{...}
}
\seealso{
Other Miscellaneous Functions: 
\code{\link{checksum}()},
\code{\link{gen.primes}()},
\code{\link{is.regex}()},
\code{\link{log_note}()},
\code{\link{polyname2orig}()},
\code{\link{unregex}()},
\code{\link{vlogical}()}
}
\concept{Miscellaneous Functions}
