% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACKAGE.R
\name{book.of.utilities}
\alias{book.of.utilities}
\title{Book of Utilities: utilities for working with data}
\description{
\code{book.of.utilities} seeks to facilitate execution of those repetitive, ad-hoc tasks often encountered during data processing.

The following functional chapters are covered in \code{book.of.utilities}:\cr
}
\section{Chapter 1 - Calculators}{

\itemize{
\item{\code{\link{calc.geo_mean}}}
\item{\code{\link{calc.harmonic_mean}}}
\item{\code{\link{calc.means}}}
\item{\code{\link{calc.rms}}}
\item{\code{\link{calc.zero_mean}}}
\item{\code{\link{range_diff}}}
\item{\code{\link{ranking.algorithm}}}
\item{\code{\link{ratio}}}
}
}

\section{Chapter 2 - Counters}{

\itemize{
\item{\code{\link{count.cycles}}}
\item{\code{\link{factor.int}}}
}
}

\section{Chapter 3 - Custom Operators}{

\itemize{
\item{\code{\link{\%?\%}}}
\item{\code{\link{\%??\%}}}
\item{\code{\link{\%><\%}}}
\item{\code{\link{\%bin\%}}}
\item{\code{\link{\%tf\%}}}
\item{\code{\link{test_between}}}
}
}

\section{Chapter 4 - Object Management}{

\itemize{
\item{\code{\link{distinct.list}}}
\item{\code{\link{enlist}}}
\item{\code{\link{get.object_sizes}}}
\item{\code{\link{scrub.data}}}
}
}

\section{Chapter 5 - Miscellaneous}{

\itemize{
\item{\code{\link{as.regex}}}
\item{\code{\link{call.recursion}}}
\item{\code{\link{checksum}}}
\item{\code{\link{gen.pass}}}
\item{\code{\link{gen.primes}}}
\item{\code{\link{is.regex}}}
\item{\code{\link{keyring_export}}}
\item{\code{\link{keyring_import}}}
\item{\code{\link{log_note}}}
\item{\code{\link{polyname2orig}}}
\item{\code{\link{radix}}}
\item{\code{\link{unregex}}}
\item{\code{\link{vlogical}}}
}
}

