% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.of.utilities.R
\name{book.of.utilities}
\alias{book.of.utilities}
\title{Book of Utilities: utilities for working with data}
\description{
\code{book.of.utilities} seeks to facilitate execution of those repetitive, ad-hoc tasks often encountered during data processing.
}
