% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculators.R
\name{calc.means}
\alias{calc.means}
\title{Calculate Means}
\usage{
calc.means(
  data,
  mean.type = "am",
  post.op = eval,
  as.zscore = FALSE,
  use.population = FALSE,
  ...
)
}
\arguments{
\item{data}{(numeric): A numeric vector}

\item{mean.type}{(string): A character vector of the types of mean value operations to execute (see "Details")}

\item{post.op}{(call | \code{\link[base]{eval}}): A function that will process the output before returning}

\item{as.zscore}{(logical | \code{FALSE}) Should the output be transformed to Z-scores?}

\item{use.population}{(logical | \code{FALSE}) Should the population standard deviation be used (ignored when \code{as.zscore==FALSE})?}

\item{...}{Additional arguments passed to other functions}
}
\value{
Arguments \code{data}, \code{mean.type}, and \code{post.op} determine the return type.
}
\description{
\code{calc.means} calculates various types of central-tendency measures for the vector passed to argument \code{a}
}
\details{
If `mean.type` contains multiple values, a list object is returned containing one member for each output; otherwise, a single value is returned with a class dependent on the value of `mean.type`.
Valid `mean.type` values are as follows:
\enumerate{
\item "am" (Arithmetic-Mean \url{https://en.wikipedia.org/wiki/Arithmetic_mean}): Return the sum of vector values divided by the magnitude of the vector
\item "zm" (Zero-Mean): Return the vector with the mean of the vector subtracted from all values
\item "gm" (Geometric Mean \url{https://en.wikipedia.org/wiki/Geometric_mean}): Return the \emph{n-th} root of the product of an \emph{n}-element input vector.  Negative input values are allowed, resulting in a complex result
\item "hm" (Harmonic Mean \url{https://en.wikipedia.org/wiki/Harmonic_mean}): Return the reciprocal of the arithmetic mean of the reciprocal of the elements of the input vector
\item "rms"(Root Mean Squared \url{https://en.wikipedia.org/wiki/Root_mean_square}): Return the square-root of the arithmetic mean of the squares of the elements of the input vector
}
}
\seealso{
Other Chapter 1 - Calculators: 
\code{\link{as.quantile}()},
\code{\link{calc.geo_mean}()},
\code{\link{calc.harmonic_mean}()},
\code{\link{calc.rms}()},
\code{\link{calc.zero_mean}()},
\code{\link{range_diff}()},
\code{\link{ranking.algorithm}()},
\code{\link{ratio}()}
}
\concept{Chapter 1 - Calculators}
