% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{call.recursion}
\alias{call.recursion}
\title{Execute a Function as Recursively}
\usage{
call.recursion(
  x,
  fun,
  test,
  ...,
  nxt = I,
  max.iter = 1,
  cur.iter = 0,
  simplify = TRUE
)
}
\arguments{
\item{x}{The input object}

\item{fun}{A function that operates on \code{x} and produces output}

\item{test}{A single-argument function returning a single Boolean: \code{FALSE} stops iteration}

\item{...}{Additional arguments passed to \code{fun}}

\item{nxt}{A function that operates on the current output of \code{fun(x)} to send to the \emph{next} iterative call to \code{fun()}}

\item{max.iter}{(integer) The maximum number of iterations}

\item{cur.iter}{(integer) The current iteration index}

\item{simplify}{(logical) Should only the last value be returned (\code{TRUE}) or intermediate values as well (\code{FALSE})?}
}
\description{
\code{call.recursion} Executes a recursive calls to \code{fun} based on the output of \code{test} up to \code{max.iter} times
}
\seealso{
Other Miscellaneous Functions: 
\code{\link{checksum}()},
\code{\link{gen.primes}()},
\code{\link{log_note}()},
\code{\link{polyname2orig}()},
\code{\link{vlogical}()}
}
\concept{Miscellaneous Functions}
