% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3. counters.R
\name{count.cycles}
\alias{count.cycles}
\title{Cycle Counter}
\usage{
count.cycles(cond, offset = 1, reset)
}
\arguments{
\item{cond}{(logical[]) A logical vector}

\item{offset}{(integer|1) An integer defining the series origin}

\item{reset}{(logical[]) A logical vector of the same length as \code{cond} such that when \code{TRUE}, the cycle counter resets to the offset}
}
\value{
A one-based (default), conditionally-incremented series
}
\description{
\code{count.cycles} detects and maps conditional occurrence frequency
}
\examples{
data.table::data.table(cond_vec = (runif(1000) >= 0.7) \%>\% sample(100), rand_col = sample(letters, 100, TRUE), key = "rand_col")[, .(cond_vec, rand_col, cyc = count.cycles(cond_vec, reset = !duplicated(rand_col)))]

}
\seealso{
Other Counters: 
\code{\link{factor.int}()}
}
\concept{Counters}
