% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2. object_management.R
\name{enlist}
\alias{enlist}
\title{Create a Named List}
\usage{
enlist(x, ...)
}
\arguments{
\item{x}{A vector of values}

\item{...}{(\code{\link[rlang]{dots_list}}):: Names to use for the list members: must be the same length as \code{x}.  Can be a vector or an atomic listing.}
}
\value{
A list, the names of which being the values of \code{x}
}
\description{
\code{enlist()} creates a list from an atomic vector.  When names having the length of the input are provided to \code{...}, a list with members named accordingly is produced.  Otherwise, the members are named by their values (eg. \code{list(this="this", that="that")} ): this is the default behavior.
}
\seealso{
Other Object management: 
\code{\link{distinct.list}()},
\code{\link{fill.na}()},
\code{\link{get.object_sizes}()},
\code{\link{scrub.data}()}
}
\concept{Object management}
