% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{gen_pass}
\alias{gen_pass}
\title{Generate a Password}
\usage{
gen_pass(exclude = "", out_len = 0L, raw = FALSE, chatty = FALSE)
}
\arguments{
\item{exclude}{Character-coercibles to exclude from output}

\item{out_len}{(integer) The length of the password in character format}

\item{raw}{(logical) Should the output be returned as raw?}

\item{chatty}{(logical) Should diagnostic information be provided?}
}
\description{
\code{gen_pass} creates a password consisting of symbols and alphanumeric values.
}
\note{
The generated string \emph{always} begins with a letter before being returned as-is or as a raw vector
}
\seealso{
Other Authentication Functions: 
\code{\link{keyring_export}()},
\code{\link{keyring_import}()},
\code{\link{kr_key}()}
}
\concept{Authentication Functions}
