% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4. custom operators.R
\name{\%all\%}
\alias{\%all\%}
\title{Retrieve all elements with the given object names}
\usage{
i \%all\% ...
}
\arguments{
\item{i}{The named input object}

\item{...}{Symbols or strings of objects to return}

\item{logical.out}{(logical) When \code{TRUE}, a logical vector or list of such is returned instead of the values satisfying the condition}

\item{chatty}{(logical) When \code{TRUE}, additional information is printed to console}
}
\value{
An object containing the values or logical results relative to the names of the input
}
\description{
This is intended to operate on a named object with duplicate names.  The idea is to allow for "soft" accessing by name
}
\seealso{
Other Custom operators: 
\code{\link{\%::\%}()},
\code{\link{\%><\%}()},
\code{\link{\%??\%}()},
\code{\link{\%?\%}()},
\code{\link{\%bin\%}()}
}
\concept{Custom operators}
