% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.of.utilities.R
\name{\%><\%}
\alias{\%><\%}
\title{"Between" Operator}
\usage{
a \%><\% z
}
\arguments{
\item{a}{Preferrably a data table or data frame, but any object that can be coerced into such}

\item{z}{Preferrably a data table or data frame, but any object that can be coerced into such}

\item{temporal}{(logical) Should relative temporal output, e.g., \code{before}, \code{during}, \code{after}, be returned as integers -1 0 1 rather than logical values?}

\item{as.text}{(logical) Should relative temporal output, e.g., \code{before}, \code{during}, \code{after}, be returned as words rather than logical values?}
}
\value{
An Boolean object comprised of the results of testing for "between-ness"
}
\description{
\code{\%><\%} chunks out each of the inputs using row-wise 'apply' and returns a boolean result matrix. When arguments \code{a} and \code{z} are vectors or tensors, "between" is interpreted as the existence of the intersection of the domains spanned by the vectors/tensors. When \code{temporal} is \code{TRUE}, an ordinal response is returned depending on the value of \code{as.text}
}
\details{
\code{as.text FALSE}: the return matrix is comprised of \code{"-1, 0, 1"} denoting \code{"before" "during" "after"}

\code{as.text TRUE}: the return matrix is comprised of \code{"before" "during" "after"}
}
\seealso{
Other Custom operators: 
\code{\link{\%::\%}()},
\code{\link{\%??\%}()},
\code{\link{\%?\%}()},
\code{\link{\%all\%}()},
\code{\link{\%bin\%}()}
}
\concept{Custom operators}
