% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_operators.R
\name{\%??\%}
\alias{\%??\%}
\title{Correlated IF-THEN-ELSE Comparison}
\usage{
cond \%??\% result
}
\arguments{
\item{cond}{(logical) A \emph{vector} or \emph{tensor} that evaluates to \code{TRUE} or \code{FALSE}}

\item{result}{(vector) Resultant values for TRUE and FALSE conditionals, ideally stored in a dimension-ed object (e.g, \code{\link[base]{data.frame}}, \code{\link[data.table]{data.table}})}
}
\description{
For each element \code{E} in \code{cond}, when \code{E} is \code{TRUE}, the corresponding index of \code{result$true} is returned: the same for \code{cond == FALSE}.
For \code{result}, the easiest way to set the available choices is to use \code{\link{\%tf\%}}; otherwise, a environment(-like) object with members named \code{true} and \code{false}
}
\seealso{
Other Custom operators: 
\code{\link{\%><\%}()},
\code{\link{\%?\%}()},
\code{\link{\%bin\%}()},
\code{\link{\%tf\%}()},
\code{\link{test_between}()}
}
\concept{Custom operators}
