% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{is.regex}
\alias{is.regex}
\title{Test for Class 'regex'}
\usage{
is.regex(i)
}
\arguments{
\item{i}{(string[]) A string or string vector}
}
\value{
Returns \code{TRUE} when the value of \code{i} contains class "regex"
}
\description{
Test for Class 'regex'
}
\seealso{
Other Miscellaneous Functions: 
\code{\link{as.regex}()},
\code{\link{checksum}()},
\code{\link{gen.primes}()},
\code{\link{log_note}()},
\code{\link{polyname2orig}()},
\code{\link{unregex}()},
\code{\link{vlogical}()}
}
\concept{Miscellaneous Functions}
