% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{keyring_import}
\alias{keyring_import}
\title{Import keyring Entries}
\usage{
keyring_import(
  data,
  kr_name = rlang::as_label(rlang::enexpr(data)),
  dry.run = FALSE
)
}
\arguments{
\item{data}{The named list of exported keyring data}

\item{kr_name}{(string) The name of the target keyring to populate: defaults to the deparsed value of argument \code{data}}

\item{dry.run}{(logical|FALSE) When \code{TRUE}, the expression that would be evaluated is returned (passwords are redacted)}
}
\value{
A logical scalar for each imported key
}
\description{
\code{keyring_import} registers exported \code{\link[keyring]{keyring}}s (see \code{\link{keyring_export}})
}
\concept{keyring Utilities}
