% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{keyring_import}
\alias{keyring_import}
\title{Import keyring Entries}
\usage{
keyring_import(
  kr_data,
  kr_name = names(kr_data),
  kr_pass = NULL,
  dry.run = FALSE
)
}
\arguments{
\item{kr_data}{The named list of exported keyring data: each name should be the name of an existing keyring.}

\item{kr_name}{(string) The names of the target keyring(s) to populate: defaults to the names of \code{kr_data}}

\item{kr_pass}{(string) The password to unlock existing keyrings:\cr
\itemize{
\item{A named list indicating the per-keyring mapping}
\item{A single password (used for all supplied keyrings)}
\item{An unnamed list the same length as \code{kr_name}}
}}

\item{dry.run}{(logical|FALSE) When \code{TRUE}, the expression that would be evaluated is returned (passwords are redacted)}
}
\value{
A list of logical vectors indicating success or failure of import for each keyring entry
}
\description{
\code{keyring_import} registers exported \code{\link[keyring]{keyring}}s (see \code{\link{keyring_export}})
}
\seealso{
Other Authentication Functions: 
\code{\link{gen.pass}()},
\code{\link{keyring_export}()},
\code{\link{kr_key}()}
}
\concept{Authentication Functions}
