% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{kr_key}
\alias{kr_key}
\title{Store an encrypted \code{keyring} key}
\usage{
kr_key(
  service = class_missing,
  username = class_missing,
  keyring = class_missing
)
}
\description{
The \code{kr_key} class
}
\section{Slots}{

\describe{
\item{\code{service,usernane,keyring}}{See \code{\link[keyring]{key_get}}}

\item{\code{key}}{A function to retrieve the \code{keyring} key}
}}

\seealso{
Other Authentication Functions: 
\code{\link{gen.pass}()},
\code{\link{keyring_export}()},
\code{\link{keyring_import}()}
}
\concept{Authentication Functions}
