% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculators.R
\name{radix}
\alias{radix}
\title{Radix Conversion}
\usage{
radix(x, ...)
}
\arguments{
\item{x}{(numeric[],symbol[]) One or more whole numbers: coercion for non-numeric and fractional input will result in loss of precision. For hexadecimal values, be sure to use the \bold{\code{"0x"}} prefix.}

\item{...}{(string,symbol,integer) The target radix to which the values in \code{x} will be converted.\cr
\itemize{
\item{Length: If two values are provided, the output is the conversion from the first radix to the second}
\item{String/symbol: If given as a strings or symbols, the following are supported (partial matching supported):\cr
\itemize{
\item{\code{binary}}
\item{\code{octal}}
\item{\code{decimal}}
\item{\code{duodecimal}}
\item{\code{hexadecimal}}
\item{\code{vigesimal}}
\item{\code{sexagesimal}}
}
}
}}
}
\value{
The \emph{decimal} representation of the input(s) using the supplied radix.
}
\description{
\code{radix} converts the \emph{representation} of the input(s) using a radix basis.  This is not to be confused with a conversion between \emph{scales}.
}
\examples{
# As binary
radix(12, bin)
radix(c(12, 20), bin)

# To hexadecimal
radix(c(12, 20), hex)

# As hex to decimal
radix(c(18, 32), hex, dec)

# Hex to binary vs. as hex to binary (equivalent outcomes)
radix(c(18, 32), hex, bin)
radix(!!radix(c(12, 20), hex), hex, bin)

# Arbitrary radix
radix(c(10, 26), 3)

# Arbitrary radix to hex
radix(c(10, 26), 3, h)

# Mixed input types to binary
radix(c(100, "0110"), b)
radix(c(100, "1001", 0x1D), b)

}
\seealso{
Other Calculators: 
\code{\link{calc.geo_mean}()},
\code{\link{calc.harmonic_mean}()},
\code{\link{calc.means}()},
\code{\link{calc.rms}()},
\code{\link{calc.zero_mean}()},
\code{\link{odds2probs}()},
\code{\link{range_diff}()},
\code{\link{ranking.algorithm}()},
\code{\link{ratio}()}
}
\concept{Calculators}
