% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculators.R
\name{ratio}
\alias{ratio}
\title{Ratio Calculator}
\usage{
ratio(i, j = i, type = NULL, decimals = 2)
}
\arguments{
\item{i}{(vector) The vector of numeric scores in the numerator}

\item{j}{(vector) The number of numeric scores in the denominator}

\item{type}{(string[]) The types of ratio algorithms to use (see Details): a vector of values is supported}

\item{decimals}{(integer | 2) The number of decimal places to which the output should be rounded}
}
\description{
\code{ratio} calculates one of the following ratio types:\cr
\enumerate{
\item{\code{"pareto"} (cumulative total vs. total)}
\item{\code{"of.max|of.min"} (relative to maximum/minimum value)}
\item{\code{NULL} (relative to total)}
}.\cr Using the related operator \code{\%ratio\%} assumes simple division by the total of \code{j}.
}
\seealso{
Other Chapter 1 - Calculators: 
\code{\link{as.quantile}()},
\code{\link{calc.geo_mean}()},
\code{\link{calc.harmonic_mean}()},
\code{\link{calc.means}()},
\code{\link{calc.rms}()},
\code{\link{calc.zero_mean}()},
\code{\link{range_diff}()},
\code{\link{ranking.algorithm}()}
}
\concept{Chapter 1 - Calculators}
