% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{vlogical}
\alias{vlogical}
\title{Vectorized Logical Tests}
\usage{
vlogical(
  vector,
  vpattern,
  test = stringi::stri_detect_regex,
  simplify_with = NULL,
  ...
)
}
\arguments{
\item{vector}{A vector or dimensional object to be processed (e.g., matrix, data.frame, etc.)}

\item{vpattern}{A \emph{named} vector of patterns to be matched}

\item{test}{(function) The function to use for logical testing: the function should be appropriate for the values of \code{vpattern} accepting \code{vector} as the first argument, the pattern as the second, and \code{...} as the third (even if not used).}

\item{simplify_with}{(function) A function to column-wise simplify the logical matrix when not empty}

\item{...}{Additional arguments to be sent to the function held by argument \code{test}}
}
\value{
A logical matrix, with rows of the same length as \code{vector} and columns the length of \code{vpattern} when \code{simplify_with} is not provided. Otherwise, the result of applying \code{simplify_with} on each column.
}
\description{
\code{vlike} is a vectorized version of \code{\link[data.table]{like}} allowing a pattern vector to be supplied
}
\examples{
	# Test 1:
vlogical(
	vector = letters
	, vpattern = c(sample(LETTERS, 5), sample(letters, 5))
	, simplify_with = mean
	, ignore.case = TRUE
	)

	# Test 2:
	vl_data <- sapply(1:100000, function(i){
				c(a = sample(LETTERS, 1), b = sample(letters, 1))
			}) |>
		data.table::as.data.table() |>
	t();

head(vl_data, 10);
dim(vl_data);

vlogical(
	vector = vl_data
	, vpattern = c(sample(LETTERS, 5), sample(letters, 5)) |> rlang::set_names()
	, test = \(vector, pattern, ...){ paste(unique(pattern), collapse = "|") |> grepl(vector)}
	, simplify_with = NULL
	, ignore.case = TRUE
	)

}
\seealso{
Other Miscellaneous Functions: 
\code{\link{as.regex}()},
\code{\link{call.recursion}()},
\code{\link{checksum}()},
\code{\link{gen.primes}()},
\code{\link{is.regex}()},
\code{\link{log_note}()},
\code{\link{polyname2orig}()},
\code{\link{unregex}()}
}
\concept{Miscellaneous Functions}
